pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Daniel Tomaszuk <dtomaszuk01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^6.2.2"
fastapi = "^0.63.0"
pyzmq = "^22.0.3"
uvicorn = "^0.13.4"
psutil = "^5.8.0"
ujson = "^4.0.2"
aiohttp = "^3.7.4"

[tool.poetry.dev-dependencies]
pre-commit = "^2.11.1"

[tool.poetry.plugins]

[tool.poetry.plugins."app.modules"]
"health" = "app.api.endpoints.health"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
exclude = '/(\.git|\.pytest_cache|docker|migrations)/'
